musicwalletproject/requirements.txt
Django==1.10.2
djangorestframework==3.5.0
mysqlclient==1.3.9
